requirements.txt
pandas
numpy
torch
transformers
sentence-transformers
faiss-cpu
scikit-learn
streamlit
accelerate
faker
matplotlib
